Improve the load process to avoid lacking of packages' error message

+++ 060_autojump.sh
@@ -1,5 +1,8 @@
 #!/bin/bash
 
-source /usr/share/autojump/autojump.bash
-export AUTOJUMP_IGNORE_CASE=1
-export AUTOJUMP_AUTOCOMPLETE_CMDS='cp vim cd'
+# Only enable autojump if existed
+if [ -f /usr/share/autojump/autojump.bash ]; then
+    source /usr/share/autojump/autojump.bash
+    export AUTOJUMP_IGNORE_CASE=1
+    export AUTOJUMP_AUTOCOMPLETE_CMDS='cp vim cd'
+fi
